change long to int in topo_info structure. This was causing some compilers to complain when print was trying to show the values using %d (oddly enough).

+++ gtopo.h
@@ -175,8 +175,8 @@ struct topo_info {
 	/* This specifies the maplet containing the above,
 	 * it changes every time we change series.
 	 */
-	long	maplet_x;
-	long	maplet_y;
+	int	maplet_x;
+	int	maplet_y;
 
 	/* fractional offset in the maplet */
 	double 	fx;
